% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeaSetClass.R
\name{brew_tea}
\alias{brew_tea}
\title{brew_tea}
\usage{
brew_tea(data = NA, ...)
}
\arguments{
\item{data}{a 3-column dataframe, defaults to NA}

\item{...}{Additional parameters that work with TeaSet$new()}
}
\value{
a TeaSet object
}
\description{
returns a TeaSet object
}
\note{
an easier way to create a TeaSet object
}
