% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeaSetClass.R
\name{extractFrom...}
\alias{extractFrom...}
\title{extractFrom...}
\usage{
extractFrom...(varNames, defaultTo = NULL, ...)
}
\arguments{
\item{varNames}{a string with the name of a variable, or a list of such strings}

\item{defaultTo}{if variable not in ..., return this.  Not designed to work with lists.}

\item{...}{Parameters from which to extract variables}
}
\value{
the variable's value, or a list of the variables' values.  NULL if cannot be found
}
\description{
Given a variable name and ..., attempts to extract the variable from ...
}
\examples{
\dontrun{extractFrom("col",...)}

}
