% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeaSetClass.R
\name{matrix_apply}
\alias{matrix_apply}
\title{matrix_apply}
\usage{
matrix_apply(x, nrow, fun, ...)
}
\arguments{
\item{x}{a 1d list of data}

\item{nrow}{how many rows to put the list into}

\item{fun}{the function to apply() to the matrix after creating it}

\item{...}{Any additional arguments for the function provided}
}
\value{
a list from using the apply function on the new matrix
}
\description{
applies the matrix function with nrow to a list before using apply with a function
}
\note{
length of new list must be evenly divisible by nrow.  This function was built to simplify some code in TeaSet
}
\examples{
test_data <- 1:16
         matrix_apply(test_data, nrow = 4, fun = sum)
}
