% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeaSetClass.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(data, type = "min-max")
}
\arguments{
\item{data}{a 2d array, matrix, or data.frame of numeric data.  Defaults to min-max}

\item{type}{a string indicating how to standardize: max, min-max, or z-score}
}
\value{
normalized data
}
\description{
normalizes or standardizes the columns of a dataframe using one of three methods: max, min-max, or zscore
}
\examples{
data(mtcars)
          normalize(mtcars, type = "z-score")
}
