% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeaSetClass.R
\name{quick_triangle}
\alias{quick_triangle}
\title{quick_triangle}
\usage{
quick_triangle(r = 1, center = c(0, 0), draw = TRUE, ...)
}
\arguments{
\item{r}{the side length}

\item{center}{the center of the triangle, as an xy-coordinate}

\item{draw}{a boolean indicating whether or not to actually draw the triangle}

\item{...}{Additional parameters that work with polygon}
}
\value{
the coordinates of the triangle's vertices, with the first coordinate relisted at the end
}
\description{
creates an equilateral triangle with all sides equal to r centered on a given point
}
\note{
not as flexible as tea_triangle() in the TeaSet class
}
