% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeaSetClass.R
\name{tea_plot}
\alias{tea_plot}
\title{tea_plot}
\usage{
tea_plot(teaSet, dataLabels = NA, main = "", newPlot = TRUE,
  bullseye = FALSE, ticks = NA, axis.labels = "default",
  col.axis = "grey", ...)
}
\arguments{
\item{teaSet}{a teaSet object. No default.}

\item{dataLabels}{labels for t-points.  If included, calls text instead of points.}

\item{main}{the title of the plot}

\item{newPlot}{whether to start a new plot, or draw atop an old plot. Defaults to TRUE}

\item{bullseye}{whether to draw a Y-shape separating i-leaf, j-leaf and k-leaf}

\item{ticks}{Whether to draw tick-lines.  Can either specify the lines (list of numerics between 0 and 1), the number of lines (integer), or whether to draw for quarters (TRUE)}

\item{axis.labels}{a 3-length list with labels for three axes.  If left as axis.labels="default", will use colnames(myCoord)}

\item{col.axis}{the color for axis lines and the bullseye.  Defaults to grey}

\item{...}{normal graphical parameters not already listed: col, pch, cex, etc.  Can also set col="contrast", "gradient", to rotate gradient, add degrees such as "gradient90"}
}
\value{
NONE
}
\description{
a plot function specifically for TeaSets.  Calls teaSet$tea_plot(parameters)
}
\note{
same functionality as teaSet$tea_plot()
}
