% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeaSetClass.R
\name{ternary_to_xy}
\alias{ternary_to_xy}
\title{ternary_to_xy}
\usage{
ternary_to_xy(data, ...)
}
\arguments{
\item{data}{data consists of a matrix or data.frame of ternary points, or a list with the coordinates for one ternary point}

\item{...}{Additional parameters that work with TeaSet$new()}
}
\value{
a matrix of xy-coordinates
}
\description{
converts ternary-coordinates to xy-coordinates
}
\note{
does not handle strings, NAs, or NULLs in data
}
